Moved plotting code to its own module and branch
